Added completion for verdi commands
This calls through the click machinery emulating a BASH terminal,
but it does not call an external subprocess so it's very fast!

I spent quite some effort to make sure it also works with weird
characters like spaces and things to escape. It should be ok now,
but there might be corner cases I didn't check. An easy way to test
is to create a computer with, e.g., a space in the name, and then
try `verdi computer show <TAB>`.
(This also fixes a partially incorrect behavior when passing parameters
to verdi, that was simply splitting on spaces).

I also fixed a raised exception when completing `{}` with missing
nodes in the history, added a newline when quitting using CTRL+D,
and added persistent history with a file that is located in the
folder of the profile (so it's different for different virtual
environments).